requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
icmplib==3.0.3
idna==3.3
plotly==5.8.0
requests==2.27.1
scapy==2.5.0
tenacity==8.0.1
urllib3==1.26.9
